fix(booking): preserve decimal deposit rates
Checkout truncated percentage deposits to whole numbers before calculating the payable amount. Preserve decimal percentage values so partial payments match the booking popup and configured rate.

+++ inc/Classes/Helper.php
@@ -3957,7 +3957,7 @@ class Helper {
 		$has_deposit = ! empty( $room['allow_deposit'] ) && $room['allow_deposit'] == true;
 		if ( $has_deposit == true ) {
 			if ( $room['deposit_type'] == 'percent' ) {
-				$deposit_amount = $price * ( intval( $room['deposit_amount'] ) / 100 );
+				$deposit_amount = $price * ( floatval( $room['deposit_amount'] ) / 100 );
 			} else {
 				$deposit_amount = $room['deposit_amount'];
 			}
